pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-project-template"
version = "1.0.0"
description = "AI-powered resume parsing and interview coaching system"
readme = "README.md"
requires-python = "==3.11.*"
license = {text = "MIT"}
authors = [
    {name = "AI Agents Team", email = "team@agent-project-template.com"}
]
keywords = ["ai", "resume", "interview", "fastapi", "pydantic-ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "fastapi==0.115.12",
    "uvicorn[standard]==0.35.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pydantic-ai[logfire]==0.8.1",
    "logfire[fastapi,httpx,redis]>=3.23.0",
    "pdfplumber==0.11.4",
    "python-multipart>=0.0.6",
    "email-validator>=2.1.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    "sonyflake-py>=1.3.0",
    "pytest-asyncio>=1.1.0",
    # NACOS configuration management
    "nacos-sdk-python>=0.1.5",
    # AWS KMS for password decryption
    "boto3>=1.39.0",
    "python-docx>=1.2.0",
    "mammoth==1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/your-org/agent-project-template"
Repository = "https://github.com/your-org/agent-project-template"
Documentation = "https://agent-project-template.readthedocs.io"
Issues = "https://github.com/your-org/agent-project-template/issues"

[project.scripts]
agent-project-template = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["ai_agents"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')"
] 

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
